040323 -3 New Features and Updated Style
Updated mat dialog box color/style.  Added feature to playlist-detail component that will load in (from user service) the user id payload information and will send the payload.user.playlists as a subject to playlist detail. used to show other playlists (from this user). THere may be an easier route to obtain this, but it works for now.